feat: change the return value of count function to long

+++ sql-genie-api/src/main/java/io/github/genie/sql/api/EntityRoot.java
@@ -40,7 +40,7 @@ public interface EntityRoot<T> {
         return get(path).avg();
     }
 
-    default NumberExpression<T, Integer> count(Path<T, ?> path) {
+    default NumberExpression<T, Long> count(Path<T, ?> path) {
         return get(path).count();
     }
